% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefma.R
\docType{package}
\name{rarefma-package}
\alias{rarefma}
\alias{rarefma-package}
\title{rarefma: PKG_TITLE}
\description{
PKG_DESC.
}
\author{
\strong{Maintainer}: AUTHOR_FIRST AUTHOR_LAST \email{AUTHOR@MAIL.COM}

}
