% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_side_performance_indicators.R
\name{sidebarIndicatorServer}
\alias{sidebarIndicatorServer}
\title{sidebar performance indicators Server Function}
\usage{
sidebarIndicatorServer(id, state)
}
\description{
sidebar performance indicators Server Function
}
